Show how to trigger actions which then trigger mutations that change the state. also show how to pass actions data
