+++ src/theme/Login.vue
@@ -2,7 +2,6 @@
   <div class="content">
     <div v-if="isAuthenticated">
       Hello Authenticated user!!!
-      <p>Name: {{profile.firstName}}</p>
       <button v-on:click="logout()" class="button is-primary">
         Logout
       </button>
@@ -55,57 +54,28 @@
 </template>
 
 <script>
-  import appService from '../app.service'
-  import { mapGetters } from 'vuex'
+  import { mapGetters, mapActions } from 'vuex'
 
   export default {
     data () {
       return {
         username: '',
-        password: '',
-        profile: {}
+        password: ''
       }
     },
     computed: {
       ...mapGetters(['isAuthenticated'])
     },
-    watch: {
-      // isAuthenticated: function (val) {
-      //   if (val) {
-      //     appService.getProfile()
-      //       .then(profile => {
-      //         this.profile = profile
-      //       })
-      //   } else {
-      //     this.profile = {}
-      //   }
-      // }
-    },
     methods: {
+      ...mapActions({
+        logout: 'logout'
+      }),
       login () {
-        appService.login({username: this.username, password: this.password})
-          .then(data => {
-            window.localStorage.setItem('token', data.token)
-            window.localStorage.setItem('tokenExpiration', data.expiration)
-            // this.isAuthenticated = true
+        this.$store.dispatch('login', {username: this.username, password: this.password})
+          .then(() => {
             this.username = ''
             this.password = ''
           })
-          .catch(() => {
-            return window.alert('Could not login')
-          })
-      },
-      logout () {
-        window.localStorage.setItem('token', null)
-        window.localStorage.setItem('tokenExpiration', null)
-        // this.isAuthenticated = false
-      }
-    },
-    created () {
-      let expiration = window.localStorage.getItem('tokenExpiration')
-      const unixTimestamp = new Date().getTime() / 1000
-      if (expiration !== null && parseInt(expiration) - unixTimestamp > 0) {
-        // this.isAuthenticated = true
       }
     }
   }
