+++ src/vuex/index.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
+import appService from '../app.service'
 
 Vue.use(Vuex)
 
@@ -17,6 +18,18 @@ const store = new Vuex.Store({
   actions: {
     logout (context) {
       context.commit('logout')
+    },
+    login (context, credantials) {
+      return new Promise(resolve => {
+        appService.login(credantials)
+          .then(data => {
+            context.commit('login', data)
+            resolve()
+          })
+          .catch(() => {
+            return window.alert('Could not login')
+          })
+      })
     }
   },
   mutations: {
@@ -26,8 +39,25 @@ const store = new Vuex.Store({
         window.localStorage.setItem('tokenExpiration', null)
       }
       state.isAuthenticated = false
+    },
+    login (state, data) {
+      if (typeof window !== 'undefined') {
+        window.localStorage.setItem('token', data.token)
+        window.localStorage.setItem('tokenExpiration', data.expiration)
+      }
+      state.isAuthenticated = true
     }
   }
 })
+
+if (typeof window !== 'undefined') {
+  document.addEventListener('DOMContentLoaded', function (event) {
+    let expiration = window.localStorage.getItem('tokenExpiration')
+    const unixTimestamp = new Date().getTime() / 1000
+    if (expiration !== null && parseInt(expiration) - unixTimestamp > 0) {
+      store.state.isAuthenticated = true
+    }
+  })
+}
 
 export default store
